refactor: slash -> application
