+++ deno/payloads/v9/interactions.ts
@@ -8,13 +8,13 @@ import type {
 	APIApplicationCommandDMInteraction,
 	APIApplicationCommandGuildInteraction,
 	APIApplicationCommandInteraction,
-} from './_interactions/slashCommands.ts';
+} from './_interactions/applicationCommands.ts';
 
 export * from './_interactions/base.ts';
 export * from './_interactions/messageComponents.ts';
 export * from './_interactions/ping.ts';
 export * from './_interactions/responses.ts';
-export * from './_interactions/slashCommands.ts';
+export * from './_interactions/applicationCommands.ts';
 
 export type APIInteraction = APIPingInteraction | APIApplicationCommandInteraction | APIMessageComponentInteraction;
