+++ payloads/v9/_interactions/applicationCommands.ts
